factor out CTA button
