+++ src/app/(public)/about/ctabutton.module.css
@@ -1,0 +1,30 @@
+.button {
+  position: relative;
+  font-size: 1.15rem;
+  color: var(--white);
+  font-weight: 700;
+  display: block;
+  width: fit-content;
+  padding-inline: 1.25em;
+  padding-block: 0.3em;
+  border-radius: 2em;
+  overflow: hidden;
+}
+
+.svg {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  inset: 0;
+  background-color: var(--color-brand-pink);
+  pointer-events: none;
+}
+
+.g1,
+.g2 {
+  transition: x 0.8s ease-in;
+  .button:hover & {
+    transition: x 0.5s ease-out;
+    x: 0%;
+  }
+}
