+++ src/app/(public)/about/whatdowedo.module.css
@@ -51,45 +51,3 @@
    * -amgg */
   box-shadow: 0px 0px 20px #a031d480;
 }
-
-/* FIXME should go somewhere shared (or into the css of a reused component mayb),
- * since this same style of button is used elsewhere too. */
-.ctaButton {
-  position: relative;
-  font-size: 1.15rem;
-  color: var(--white);
-  font-weight: 700;
-  display: block;
-  width: fit-content;
-  padding-inline: 1.25em;
-  padding-block: 0.3em;
-  border-radius: 2em;
-  /*
-  background: linear-gradient(
-    90deg in oklab,
-    var(--blurple),
-    var(--color-brand-pink),
-    var(--yellow)
-  );
-  background-position: center;
-  background-size: 300% 100%;
-  transition: background-size .5s ease;
-  &:hover {
-    background-size: 100% 100%;
-  }
-  */
-  background: linear-gradient(90deg in oklab, var(--blurple), var(--color-brand-pink), var(--yellow));
-}
-
-.ctaButton > svg {
-  pointer-events: none;
-}
-
-.g1,
-.g2 {
-  transition: x 0.8s ease-in;
-  .ctaButton:hover & {
-    transition: x 0.5s ease-out;
-    x: 0%;
-  }
-}
